got visuals working, got commands to sync
